refactor: change output 분리
