+++ src/components/Change/index.tsx
@@ -1,0 +1,13 @@
+import ChangeInput from 'components/ChangeInput';
+import ChangeOutput from 'components/ChangeOutput';
+
+function Change() {
+  return (
+    <>
+      <ChangeInput />
+      <ChangeOutput />
+    </>
+  );
+}
+
+export default Change;
